+++ packages/eyes-sdk-core/lib/sdk/CheckSettingsUtils.js
@@ -97,6 +97,7 @@ function toCheckWindowConfiguration({checkSettings, configuration}) {
     enablePatterns: TypeUtils.getOrDefault(checkSettings.enablePatterns, configuration.getEnablePatterns()),
     useDom: TypeUtils.getOrDefault(checkSettings.useDom, configuration.getUseDom()),
     variationGroupId: checkSettings.variationGroupId,
+    pageId: checkSettings.pageId || undefined,
   }
 
   if (config.target === 'region') {
@@ -206,12 +207,17 @@ async function toScreenshotCheckSettings({checkSettings, context, screenshot}) {
           const contextLocationInViewport = await elements[0].context.getLocationInViewport()
           for (const element of elements) {
             const region = utils.geometry.offset(await element.getRegion(), contextLocationInViewport)
-            referenceRegions.push({
-              x: Math.max(0, region.x - screenshot.region.x),
-              y: Math.max(0, region.y - screenshot.region.y),
-              width: region.width,
-              height: region.height,
-            })
+            referenceRegions.push(
+              utils.geometry.scale(
+                {
+                  x: Math.max(0, region.x - screenshot.region.x),
+                  y: Math.max(0, region.y - screenshot.region.y),
+                  width: region.width,
+                  height: region.height,
+                },
+                context.driver.viewportScale,
+              ),
+            )
           }
         }
       }