+++ packages/visual-grid-client/src/sdk/checkWindow.js
@@ -51,6 +51,7 @@ function makeCheckWindow({
     enablePatterns,
     ignoreDisplacements,
     visualGridOptions = _visualGridOptions,
+    pageId,
     closeAfterMatch,
     throwEx = true,
     variationGroupId,
@@ -179,6 +180,7 @@ function makeCheckWindow({
         scriptHooks,
         sendDom,
         visualGridOptions,
+        includeFullPageSize: !!pageId,
       })
 
       if (!wrapper.getAppEnvironment()) {
@@ -240,6 +242,7 @@ function makeCheckWindow({
         imageLocation: screenshotUrl,
         domLocation,
         selectorRegions,
+        fullPageSize,
         imagePositionInActiveFrame: imageLocation,
       } = renderStatusResult
 
@@ -286,7 +289,6 @@ function makeCheckWindow({
       }
 
       logger.verbose(`running wrapper.checkWindow for test ${testName} stepCount #${currStepCount}`)
-
       const checkArgs = {
         screenshotUrl,
         tag,
@@ -296,6 +298,13 @@ function makeCheckWindow({
         url,
         closeAfterMatch,
         throwEx,
+      }
+      if (pageId) {
+        checkArgs.pageCoverageInfo = {
+          pageId,
+          ...fullPageSize,
+          imagePositionInPage: {x: imageLocation.x, y: imageLocation.y},
+        }
       }
 
       return wrapper.checkWindow(checkArgs)