+++ packages/types/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Unreleased
 
 
+## 1.2.0 - 2022/3/6
+
+- Support 'AUTProxy' capability
+
 ## 1.1.1 - 2022/3/3
 
 - fix `pageId` optional property